# Improve cache performance for bare string values