--- a/activesupport/test/cache/serializer_with_fallback_test.rb
+++ b/activesupport/test/cache/serializer_with_fallback_test.rb
@@ -45,6 +45,64 @@ class CacheSerializerWithFallbackTest < ActiveSupport::TestCase
     end
   end
 
+  (FORMATS - [:passthrough, :marshal_6_1, :marshal_7_0]).each do |format|
+    test "#{format.inspect} serializer preserves version with bare string" do
+      entry = ActiveSupport::Cache::Entry.new("abc", version: "123")
+      assert_entry entry, roundtrip(format, entry)
+    end
+
+    test "#{format.inspect} serializer preserves expiration with bare string" do
+      entry = ActiveSupport::Cache::Entry.new("abc", expires_in: 123)
+      assert_entry entry, roundtrip(format, entry)
+    end
+
+    test "#{format.inspect} serializer preserves encoding of version with bare string" do
+      [Encoding::UTF_8, Encoding::BINARY].each do |encoding|
+        version = "123".encode(encoding)
+        roundtripped = roundtrip(format, ActiveSupport::Cache::Entry.new("abc", version: version))
+        assert_equal version.encoding, roundtripped.version.encoding
+      end
+    end
+
+    test "#{format.inspect} serializer preserves encoding of bare string" do
+      [Encoding::UTF_8, Encoding::BINARY, Encoding::US_ASCII].each do |encoding|
+        string = "abc".encode(encoding)
+        roundtripped = roundtrip(format, ActiveSupport::Cache::Entry.new(string))
+        assert_equal string.encoding, roundtripped.value.encoding
+      end
+    end
+
+    test "#{format.inspect} serializer dumps bare string with reduced overhead when possible" do
+      string = "abc"
+      options = { version: "123", expires_in: 123 }
+
+      unsupported = string.encode(Encoding::WINDOWS_1252)
+      unoptimized = serializer(format).dump(ActiveSupport::Cache::Entry.new(unsupported, **options))
+
+      [Encoding::UTF_8, Encoding::BINARY, Encoding::US_ASCII].each do |encoding|
+        supported = string.encode(encoding)
+        optimized = serializer(format).dump(ActiveSupport::Cache::Entry.new(supported, **options))
+        assert_operator optimized.size, :<, unoptimized.size
+      end
+    end
+
+    test "#{format.inspect} serializer can compress bare strings" do
+      entry = ActiveSupport::Cache::Entry.new("abc" * 100, version: "123", expires_in: 123)
+      compressed = serializer(format).dump_compressed(entry, 1)
+      uncompressed = serializer(format).dump_compressed(entry, 100_000)
+      assert_operator compressed.bytesize, :<, uncompressed.bytesize
+    end
+  end
+
+  [:passthrough, :marshal_6_1, :marshal_7_0].each do |format|
+    test "#{format.inspect} serializer dumps bare string in a backward compatible way" do
+      string = +"abc"
+      string.instance_variable_set(:@baz, true)
+      roundtripped = roundtrip(format, ActiveSupport::Cache::Entry.new(string))
+      assert roundtripped.value.instance_variable_get(:@baz)
+    end
+  end
+
   test ":message_pack serializer handles missing class gracefully" do
     klass = Class.new do
       def self.name; "DoesNotActuallyExist"; end
@@ -68,10 +126,14 @@ class CacheSerializerWithFallbackTest < ActiveSupport::TestCase
       ActiveSupport::Cache::SerializerWithFallback[format]
     end
 
+    def roundtrip(format, entry)
+      serializer(format).load(serializer(format).dump(entry))
+    end
+
     def assert_entry(expected, actual)
-      assert_equal expected.value, actual.value
-      assert_equal expected.version, actual.version
-      assert_equal expected.expires_at, actual.expires_at
+      assert_equal \
+        [expected.value, expected.version, expected.expires_at],
+        [actual.value, actual.version, actual.expires_at]
     end
 
     def assert_logs(pattern, &block)