@sliumdx 2025-10-30
> (A new cache format is necessary because, during a rolling deploy, unupgraded servers must be able to read cache entries from upgraded servers, which means the optimization cannot be enabled for existing apps by default.)

Ran into this issue when upgrading rails from 7.0 to 7.1, do you think we could update the upgrade guide(like how that is done for upgrading 6.1-> 7) as mentioned in  https://github.com/rails/rails/pull/42076

CC: @byroot  @jonathanhefner